Added State type deserialization
State can be bools, ints, floats, string or null. Because HomeAssistant does
not do a good job of keeping track of types (some types are serialized
as strings instead of their underlying type), I wrote a visitor that will
attempt to parse out the underlying type. This is very inefficent since
we have to try to parse out each type it could be.  Regardless I still think
it is worth doing since the user will have todo it anyway.