Added serde_debugging feature (#12)
Added the serde_debugging feature which will make debugging serde
deserialization errors easier. This feature is disabled by default since
there is a cost in using it and in general should not be used in
production enviroments. Also updated the CI which will now build/lint
both the default build and the one with the `serde_debugging` feature
enabled. Will update our test cases with the debugging feature and create
a live endpoint test soon.